Changing this in hopes this won't break anything. (It won't right, Ninji?)

+++ polaris_shared/src/packets/Packet.h
@@ -9,7 +9,7 @@ struct PacketHeader {
     uint8_t subcommand;
     uint8_t flag1;
     uint8_t flag2;
-    PacketHeader(uint32_t length, uint8_t command, uint8_t subcommand, uint8_t flag1, uint8_t flag2) {
+    PacketHeader(size_t length, uint8_t command, uint8_t subcommand, uint8_t flag1, uint8_t flag2) {
         this->length = length;
         this->command = command;
         this->subcommand = subcommand;
